requirements.txt
Flask
Flask-Login
Flask-Bcrypt
Flask-SQLAlchemy
psycopg2
python-dotenv
matplotlib
requests